Done : Exercises 5.5.-5.11.
5.5 Blog list frontend, step5
5.6 Blog list frontend, step6
5.7 Blog list frontend, step7
5.8: Blog list frontend, step8
5.9: Blog list frontend, step9
5.10: Blog list frontend, step10
5.11: Blog list frontend, step11

+++ part4/blog_list/controllers/blogs.js
@@ -18,6 +18,7 @@ Exercises 4.15.-4.23.
 
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
 const { tokenExtractor, userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (req, res) => {
@@ -36,36 +37,57 @@ blogsRouter.post("/", tokenExtractor, userExtractor, async (req, res) => {
   const { user } = req;
   blog.user = user._id;
   const blogToSave = new Blog(blog);
-
   const savedBlog = await blogToSave.save();
 
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
 
-  res.status(201).json(savedBlog);
+  const mergedBlog = await savedBlog.populate("user", { username: 1, name: 1 });
+  res.status(201).json(mergedBlog);
 });
 
 blogsRouter.put("/:id", async (req, res) => {
+  const blogToUpdate = await Blog.findById(req.params.id);
   const blog = {
     title: req.body.title,
     author: req.body.author,
     url: req.body.url,
     likes: req.body.likes,
+    user: req.body.user,
   };
   const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blog, {
     new: true,
   });
+  if(blogToUpdate.user !== req.body.user){
+    const user = await User.findById(req.body.user);
+    console.log(user);
+    user.blogs = user.blogs.concat(updatedBlog._id);
+    await user.save();
+    
+    const blogToRemoveFromUser = await User.findById(blogToUpdate.user);
+    blogToRemoveFromUser.blogs.pop(blogToUpdate.id);
+    await blogToRemoveFromUser.save();
+
+  }
   res.json(updatedBlog);
+  
 });
 
 blogsRouter.delete("/:id", tokenExtractor, userExtractor, async (req, res) => {
   const { user } = req;
   if (user) {
     const blogToRemove = await Blog.findById(req.params.id);
-
+    console.log(blogToRemove);
+    if (blogToRemove.user === undefined) {
+      return res.status(401).json({ error: "unauthorised" });
+    }
     if (user.id.toString() === blogToRemove.user.toString()) {
-      const result = await Blog.findByIdAndRemove(req.params.id);
-      console.log(result);
+      const deletedBlog = await Blog.findByIdAndRemove(req.params.id);
+      if(deletedBlog){
+        const blogToRemoveFromUser = await User.findById(deletedBlog.user);
+        blogToRemoveFromUser.blogs.pop(deletedBlog.id);
+        await blogToRemoveFromUser.save();
+      }
       res.status(204).end();
     } else {
       return res.status(401).json({ error: "unauthorised" });
